Backend Server/esp_flashing_tool/requirements.txt
fastapi
uvicorn
pyserial
qrcode
Pillow
jinja2 